Fix BusPirate.response() regression from API change
Commit 4f8ae283e87b79706a726b84485cd75e650d4c8d changed the definition
of BBIO_Base.response() (BBIO_Base was later renamed to BusPirate) such
that the second argument went from a boolean that indicates that data
should be returned to a boolean that indicates if the data should be
decoded. There were a few cases where the old meaning was still assumed.

To prevent a similar bug, all instances of response() with a second
argument now use a keyword argument.